@@ -5,21 +5,22 @@
   "private": true,
   "dependencies": {
     "@paypal/react-paypal-js": "^7.8.2",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.24.0",
-    "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.1.0",
-    "react-dom": "^17.0.2",
-    "react-helmet-async": "^1.2.2",
-    "react-router-bootstrap": "^0.26.0",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.3.4",
+    "bootstrap": "^5.2.3",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.2",
+    "react-dom": "^18.2.0",
+    "react-google-charts": "^4.0.0",
+    "react-helmet-async": "^1.3.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.9.0",
+    "react-scripts": "5.0.1",
     "react-toastify": "^9.1.1",
     "use-reducer-logger": "^1.0.2",
-    "web-vitals": "^2.1.2"
+    "web-vitals": "^3.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
