@@ -1,5 +1,6 @@
-function Rating(props) {
-  const { rating, numReviews } = props;
+function Rating(props)
+{
+  const { rating, numReviews, caption } = props;
   return (
     <div className="rating">
       <span>
@@ -8,8 +9,8 @@ function Rating(props) {
             rating >= 1
               ? 'fas fa-star'
               : rating >= 0.5
-              ? 'fas fa-star-half-alt'
-              : 'far fa-star'
+                ? 'fas fa-star-half-alt'
+                : 'far fa-star'
           }
         />
       </span>
@@ -19,8 +20,8 @@ function Rating(props) {
             rating >= 2
               ? 'fas fa-star'
               : rating >= 1.5
-              ? 'fas fa-star-half-alt'
-              : 'far fa-star'
+                ? 'fas fa-star-half-alt'
+                : 'far fa-star'
           }
         />
       </span>
@@ -30,8 +31,8 @@ function Rating(props) {
             rating >= 3
               ? 'fas fa-star'
               : rating >= 2.5
-              ? 'fas fa-star-half-alt'
-              : 'far fa-star'
+                ? 'fas fa-star-half-alt'
+                : 'far fa-star'
           }
         />
       </span>
@@ -41,8 +42,8 @@ function Rating(props) {
             rating >= 4
               ? 'fas fa-star'
               : rating >= 3.5
-              ? 'fas fa-star-half-alt'
-              : 'far fa-star'
+                ? 'fas fa-star-half-alt'
+                : 'far fa-star'
           }
         />
       </span>
@@ -52,12 +53,16 @@ function Rating(props) {
             rating >= 5
               ? 'fas fa-star'
               : rating >= 4.5
-              ? 'fas fa-star-half-alt'
-              : 'far fa-star'
+                ? 'fas fa-star-half-alt'
+                : 'far fa-star'
           }
         />
       </span>
-      <span> {numReviews} reviews</span>
+      {caption ? (
+        <span>{caption}</span>
+      ) : (
+        <span>{' ' + numReviews + ' reviews'}</span>
+      )}
     </div>
   );
 }
