@@ -5,6 +5,7 @@ import { Helmet } from 'react-helmet-async';
 import { useNavigate, useParams } from 'react-router-dom';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
+import Button from 'react-bootstrap/Button';
 import ListGroup from 'react-bootstrap/ListGroup';
 import Card from 'react-bootstrap/Card';
 import { Link } from 'react-router-dom';
@@ -32,6 +33,18 @@ function reducer(state, action)
             return { ...state, loadingPay: false };
         case 'PAY_RESET':
             return { ...state, loadingPay: false, successPay: false };
+        case 'DELIVER_REQUEST':
+            return { ...state, loadingDeliver: true };
+        case 'DELIVER_SUCCESS':
+            return { ...state, loadingDeliver: false, successDeliver: true };
+        case 'DELIVER_FAIL':
+            return { ...state, loadingDeliver: false };
+        case 'DELIVER_RESET':
+            return {
+                ...state,
+                loadingDeliver: false,
+                successDeliver: false,
+            };
         default:
             return state;
     }
@@ -45,14 +58,24 @@ export default function OrderScreen()
     const { id: orderId } = params;
     const navigate = useNavigate();
 
-    const [{ loading, error, order, successPay, loadingPay }, dispatch] =
-        useReducer(reducer, {
-            loading: true,
-            order: {},
-            error: '',
-            successPay: false,
-            loadingPay: false,
-        });
+    const [
+        {
+            loading,
+            error,
+            order,
+            successPay,
+            loadingPay,
+            loadingDeliver,
+            successDeliver,
+        },
+        dispatch,
+    ] = useReducer(reducer, {
+        loading: true,
+        order: {},
+        error: '',
+        successPay: false,
+        loadingPay: false,
+    });
 
     const [{ isPending }, paypalDispatch] = usePayPalScriptReducer();
 
@@ -121,12 +144,21 @@ export default function OrderScreen()
         {
             return navigate('/login');
         }
-        if (!order._id || successPay || (order._id && order._id !== orderId))
+        if (
+            !order._id ||
+            successPay ||
+            successDeliver ||
+            (order._id && order._id !== orderId)
+        )
         {
             fetchOrder();
             if (successPay)
             {
                 dispatch({ type: 'PAY_RESET' });
+            }
+            if (successDeliver)
+            {
+                dispatch({ type: 'DELIVER_RESET' });
             }
         } else
         {
@@ -146,7 +178,36 @@ export default function OrderScreen()
             };
             loadPaypalScript();
         }
-    }, [order, userInfo, orderId, navigate, paypalDispatch, successPay]);
+    }, [
+        order,
+        userInfo,
+        orderId,
+        navigate,
+        paypalDispatch,
+        successPay,
+        successDeliver,
+    ]);
+
+    async function deliverOrderHandler()
+    {
+        try
+        {
+            dispatch({ type: 'DELIVER_REQUEST' });
+            const { data } = await axios.put(
+                `/api/orders/${order._id}/deliver`,
+                {},
+                {
+                    headers: { authorization: `Bearer ${userInfo.token}` },
+                }
+            );
+            dispatch({ type: 'DELIVER_SUCCESS', payload: data });
+            toast.success('Order is delivered');
+        } catch (err)
+        {
+            toast.error(getError(err));
+            dispatch({ type: 'DELIVER_FAIL' });
+        }
+    }
     return loading ? (
         <LoadingBox></LoadingBox>
     ) : error ? (
@@ -266,6 +327,16 @@ export default function OrderScreen()
                                             </div>
                                         )}
                                         {loadingPay && <LoadingBox></LoadingBox>}
+                                    </ListGroup.Item>
+                                )}
+                                {userInfo.isAdmin && order.isPaid && !order.isDelivered && (
+                                    <ListGroup.Item>
+                                        {loadingDeliver && <LoadingBox></LoadingBox>}
+                                        <div className="d-grid">
+                                            <Button type="button" onClick={deliverOrderHandler}>
+                                                Deliver Order
+                                            </Button>
+                                        </div>
                                     </ListGroup.Item>
                                 )}
                             </ListGroup>
