@@ -1,10 +1,11 @@
 import axios from 'axios';
-import { useContext, useEffect, useReducer } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useContext, useEffect, useReducer, useRef, useState } from 'react';
+import { Link, useNavigate, useParams } from 'react-router-dom';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Card from 'react-bootstrap/Card';
 import ListGroup from 'react-bootstrap/ListGroup';
+import Form from 'react-bootstrap/Form';
 import Badge from 'react-bootstrap/Badge';
 import Button from 'react-bootstrap/Button';
 import Rating from '../components/Rating';
@@ -13,9 +14,21 @@ import LoadingBox from '../components/LoadingBox';
 import MessageBox from '../components/MessageBox';
 import { getError } from '../utils';
 import { Store } from '../Store';
+import FloatingLabel from 'react-bootstrap/FloatingLabel';
+import { toast } from 'react-toastify';
 
-const reducer = (state, action) => {
-  switch (action.type) {
+const reducer = (state, action) =>
+{
+  switch (action.type)
+  {
+    case 'REFRESH_PRODUCT':
+      return { ...state, product: action.payload };
+    case 'CREATE_REQUEST':
+      return { ...state, loadingCreateReview: true };
+    case 'CREATE_SUCCESS':
+      return { ...state, loadingCreateReview: false };
+    case 'CREATE_FAIL':
+      return { ...state, loadingCreateReview: false };
     case 'FETCH_REQUEST':
       return { ...state, loading: true };
     case 'FETCH_SUCCESS':
@@ -27,23 +40,35 @@ const reducer = (state, action) => {
   }
 };
 
-function ProductScreen() {
+function ProductScreen()
+{
+  let reviewsRef = useRef();
+
+  const [rating, setRating] = useState(0);
+  const [comment, setComment] = useState('');
+  const [selectedImage, setSelectedImage] = useState('');
   const navigate = useNavigate();
   const params = useParams();
   const { slug } = params;
 
-  const [{ loading, error, product }, dispatch] = useReducer(reducer, {
-    product: [],
-    loading: true,
-    error: '',
-  });
-  useEffect(() => {
-    const fetchData = async () => {
+  const [{ loading, error, product, loadingCreateReview }, dispatch] =
+    useReducer(reducer, {
+      product: [],
+      loading: true,
+      error: '',
+    });
+
+  useEffect(() =>
+  {
+    const fetchData = async () =>
+    {
       dispatch({ type: 'FETCH_REQUEST' });
-      try {
+      try
+      {
         const result = await axios.get(`/api/products/slug/${slug}`);
         dispatch({ type: 'FETCH_SUCCESS', payload: result.data });
-      } catch (err) {
+      } catch (err)
+      {
         dispatch({ type: 'FETCH_FAIL', payload: getError(err) });
       }
     };
@@ -51,12 +76,14 @@ function ProductScreen() {
   }, [slug]);
 
   const { state, dispatch: ctxDispatch } = useContext(Store);
-  const { cart } = state;
-  const addToCartHandler = async () => {
+  const { cart, userInfo } = state;
+  const addToCartHandler = async () =>
+  {
     const existItem = cart.cartItems.find((x) => x._id === product._id);
     const quantity = existItem ? existItem.quantity + 1 : 1;
     const { data } = await axios.get(`/api/products/${product._id}`);
-    if (data.countInStock < quantity) {
+    if (data.countInStock < quantity)
+    {
       window.alert('Sorry. Product is out of stock');
       return;
     }
@@ -66,6 +93,44 @@ function ProductScreen() {
     });
     navigate('/cart');
   };
+
+  const submitHandler = async (e) =>
+  {
+    e.preventDefault();
+    if (!comment || !rating)
+    {
+      toast.error('Please enter comment and rating');
+      return;
+    }
+    try
+    {
+      const { data } = await axios.post(
+        `/api/products/${product._id}/reviews`,
+        { rating, comment, name: userInfo.name },
+        {
+          headers: { Authorization: `Bearer ${userInfo.token}` },
+        }
+      );
+
+      dispatch({
+        type: 'CREATE_SUCCESS',
+      });
+      toast.success('Review submitted successfully');
+      product.reviews.unshift(data.review);
+      product.numReviews = data.numReviews;
+      product.rating = data.rating;
+      dispatch({ type: 'REFRESH_PRODUCT', payload: product });
+      window.scrollTo({
+        behavior: 'smooth',
+        top: reviewsRef.current.offsetTop,
+      });
+    } catch (error)
+    {
+      toast.error(getError(error));
+      dispatch({ type: 'CREATE_FAIL' });
+    }
+  };
+
   return loading ? (
     <LoadingBox />
   ) : error ? (
@@ -76,7 +141,7 @@ function ProductScreen() {
         <Col md={6}>
           <img
             className="img-large"
-            src={product.image}
+            src={selectedImage || product.image}
             alt={product.name}
           ></img>
         </Col>
@@ -95,6 +160,24 @@ function ProductScreen() {
               ></Rating>
             </ListGroup.Item>
             <ListGroup.Item>Pirce : ${product.price}</ListGroup.Item>
+            <ListGroup.Item>
+              <Row xs={1} md={2} className="g-2">
+                {[product.image, ...product.images].map((x) => (
+                  <Col key={x}>
+                    <Card>
+                      <Button
+                        className="thumbnail"
+                        type="button"
+                        variant="light"
+                        onClick={() => setSelectedImage(x)}
+                      >
+                        <Card.Img variant="top" src={x} alt="product" />
+                      </Button>
+                    </Card>
+                  </Col>
+                ))}
+              </Row>
+            </ListGroup.Item>
             <ListGroup.Item>
               Description:
               <p>{product.description}</p>
@@ -138,6 +221,73 @@ function ProductScreen() {
           </Card>
         </Col>
       </Row>
+      <div className="my-3">
+        <h2 ref={reviewsRef}>Reviews</h2>
+        <div className="mb-3">
+          {product.reviews.length === 0 && (
+            <MessageBox>There is no review</MessageBox>
+          )}
+        </div>
+        <ListGroup>
+          {product.reviews.map((review) => (
+            <ListGroup.Item key={review._id}>
+              <strong>{review.name}</strong>
+              <Rating rating={review.rating} caption=" "></Rating>
+              <p>{review.createdAt.substring(0, 10)}</p>
+              <p>{review.comment}</p>
+            </ListGroup.Item>
+          ))}
+        </ListGroup>
+        <div className="my-3">
+          {userInfo ? (
+            <form onSubmit={submitHandler}>
+              <h2>Write a customer review</h2>
+              <Form.Group className="mb-3" controlId="rating">
+                <Form.Label>Rating</Form.Label>
+                <Form.Select
+                  aria-label="Rating"
+                  value={rating}
+                  onChange={(e) => setRating(e.target.value)}
+                >
+                  <option value="">Select...</option>
+                  <option value="1">1- Poor</option>
+                  <option value="2">2- Fair</option>
+                  <option value="3">3- Good</option>
+                  <option value="4">4- Very good</option>
+                  <option value="5">5- Excelent</option>
+                </Form.Select>
+              </Form.Group>
+              <FloatingLabel
+                controlId="floatingTextarea"
+                label="Comments"
+                className="mb-3"
+              >
+                <Form.Control
+                  as="textarea"
+                  placeholder="Leave a comment here"
+                  value={comment}
+                  onChange={(e) => setComment(e.target.value)}
+                />
+              </FloatingLabel>
+
+              <div className="mb-3">
+                <Button disabled={loadingCreateReview} type="submit">
+                  Submit
+                </Button>
+                {loadingCreateReview && <LoadingBox></LoadingBox>}
+              </div>
+            </form>
+          ) : (
+            <MessageBox>
+              Please{' '}
+              <Link to={`/signin?redirect=/product/${product.slug}`}>
+                Sign In
+              </Link>{' '}
+              to write a review
+            </MessageBox>
+          )}
+        </div>
+      </div>
     </div>
   );
 }
